backend/requirements/base.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-otp==1.2.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7

# Cache & Performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# File Storage
boto3==1.29.7
django-storages==1.14.2
Pillow==10.1.0

# API Documentation
drf-spectacular==0.26.5

# Monitoring & Logging
sentry-sdk==1.38.0

# ML & Analytics
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4

# Payment Integrations
requests==2.31.0

# Export & Reports
openpyxl==3.1.2
reportlab==4.0.7

# Background Tasks
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Additional dependencies
whitenoise==6.6.0
gunicorn==21.2.0